+++ week4/junghyun/mission/src/components/FloatingButton.jsx
@@ -1,0 +1,10 @@
+import icon_plus from "../assets/plus.svg";
+
+const FloatingButton = () => {
+  return (
+    <button class="write__btn" type="button">
+      <img src={icon_plus} alt="write button" />
+    </button> 
+  );
+};
+export default FloatingButton;
